impove parallel run and status display.

+++ cvsTool/Model/Simulator.cs
@@ -85,7 +85,7 @@ namespace cvsTool.Model
 
         public void startSimulate_bak()
         {
-            for (int i = 1; i < 10; i++)
+            for (UInt16 i = 1; i < 10; i++)
             {
                 for (int j = 1; j < 10; j++)
                 {
